LYNX-760: initial implementation
