+++ .eslintrc.js
@@ -15,12 +15,12 @@ module.exports = {
     'linebreak-style': ['error', 'unix'], // enforce unix linebreaks
     'no-param-reassign': [2, { props: false }], // allow modifying properties of param
     'no-use-before-define': [2, { functions: false }],
-    'no-console': [
-      'error',
-      {
-        allow: ['warn', 'error', 'info', 'debug'],
-      },
-    ],
+    // 'no-console': [
+    //   'error',
+    //   {
+    //     allow: ['warn', 'error', 'info', 'debug'],
+    //   },
+    // ],
     'no-unused-vars': ['error', { argsIgnorePattern: '^_', varsIgnorePattern: '^_' }],
   },
 };
