+++ blocks/header/menu/hashTagParser.js
@@ -1,0 +1,39 @@
+/**
+ * Parses href hash
+ *
+ * @param urlHash
+ * @returns {*[]}
+ */
+const namespaces = [
+  'display_for_',
+  'apply_style_',
+];
+
+function parseUrlHashTag(href) {
+  if (!href.indexOf('#') > 0) {
+    return;
+  }
+
+  const parsed = [];
+  const hashTags = href.split('#').slice(1);
+
+  if (hashTags) {
+    namespaces.forEach((ns) => {
+      hashTags.forEach((tag) => {
+        const value = tag.split(ns);
+        if (value && value.length === 2) {
+          parsed.push({
+            namespace: ns,
+            value: value[1],
+          });
+        }
+      });
+    });
+  }
+
+  return parsed;
+}
+
+export {
+  parseUrlHashTag,
+};
